+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This change log is started in 0.3.2;
 
 ## 0.3.4 (2016-04-28)
 - Using algorithm and password in pvDecrypt
+- Using algorithm and password in pvEncrypt
 
 ## 0.3.3 (2016-04-26)
 - Show error if password is invalid - review